암호화분리&상수화처리
