+++ src/main/java/flab/delideli/encrypt/EncryptPassword.java
@@ -1,0 +1,25 @@
+package flab.delideli.encrypt;
+
+import org.springframework.stereotype.Component;
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+@Component
+public class EncryptPassword {
+
+    public String encrypt(String password) throws NoSuchAlgorithmException {
+        MessageDigest md = MessageDigest.getInstance("SHA-256");
+        md.update(password.getBytes());
+
+        return bytesToHex(md.digest());
+    }
+
+    private String bytesToHex(byte[] bytes) {
+        StringBuilder builder = new StringBuilder();
+        for(byte b: bytes) {
+            builder.append(String.format("%02x",b));
+        }
+        return builder.toString();
+    }
+}
