+++ src/main/java/flab/delideli/service/MemberServiceImpl.java
@@ -2,6 +2,7 @@ package flab.delideli.service;
 
 import flab.delideli.dao.MemberDao;
 import flab.delideli.dto.MemberDTO;
+import flab.delideli.encrypt.EncryptPassword;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,12 +14,13 @@ import java.security.NoSuchAlgorithmException;
 public class MemberServiceImpl implements MemberService {
 
     private final MemberDao memberDao;
+    private final EncryptPassword encryptPassword;
 
     @Override
     public void joinMember(MemberDTO member) throws NoSuchAlgorithmException {
         if(this.isExistUserId(member.getUserid()))
             throw new IllegalArgumentException("이미 존재하는 아이디입니다.");
-        MemberDTO encodemember = new MemberDTO(member.getUserid(), encrypt(member.getPassword()), member.getUsername(), member.getPhone(), member.getAddress());
+        MemberDTO encodemember = new MemberDTO(member.getUserid(), encryptPassword.encrypt(member.getPassword()), member.getUsername(), member.getPhone(), member.getAddress());
         memberDao.joinMember(encodemember);
     }
 
@@ -27,19 +29,4 @@ public class MemberServiceImpl implements MemberService {
         return memberDao.isExistUserId(userid);
     }
 
-
-    public String encrypt(String password) throws NoSuchAlgorithmException {
-        MessageDigest md = MessageDigest.getInstance("SHA-256");
-        md.update(password.getBytes());
-
-        return bytesToHex(md.digest());
-    }
-
-    private String bytesToHex(byte[] bytes) {
-        StringBuilder builder = new StringBuilder();
-        for(byte b: bytes) {
-            builder.append(String.format("%02x",b));
-        }
-        return builder.toString();
-    }
 }
